added modules, more controls, navigation buttons and instructions
